Optimized Code! - Optimized remaining C++ code - Added Pfit - Updated examples to use erupt()
