+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: MultiscaleDTM
 Title: Multi-Scale Geomorphometric Terrain Attributes
-Version: 0.9.3
+Version: 1.0
 Authors@R: 
      c(
     person(given = "Alexander",
