+++ NAMESPACE
@@ -1,21 +1,18 @@
 # Generated by roxygen2: do not edit by hand
 
 export(AdjSD)
-export(AdjSD_arma)
 export(BPI)
 export(DMV)
 export(DirSlp)
+export(Pfit)
 export(Qfit)
-export(Qfit_arma)
 export(RIE)
 export(RelPos)
 export(SAPA)
 export(SlpAsp)
 export(SurfaceArea)
-export(SurfaceArea_optimized)
 export(TPI)
 export(VRM)
-export(VRM_optimized)
 export(annulus_window)
 export(circle_window)
 export(erupt)
