+++ R/DMV.R
@@ -13,9 +13,7 @@
 #' @param wopt List with named options for writing files as in writeRaster.
 #' @return a SpatRaster or RasterLayer.
 #' @examples
-#' r<- rast(volcano, extent= ext(2667400, 2667400 + 
-#' ncol(volcano)*10, 6478700, 6478700 + nrow(volcano)*10), 
-#' crs = "EPSG:27200")
+#' r<- erupt()
 #' dmv<- DMV(r, w=c(5,5), shape= "rectangle", stand="range", na.rm = TRUE)
 #' plot(dmv)
 #' @import terra
