+++ R/DirSlope.R
@@ -19,9 +19,11 @@
 #' r<- erupt()
 #' dz1<- SlpAsp(r, metrics = c("dz.dx", "dz.dy"))
 #' dz2<- Qfit(r, metrics = c(), return_params = TRUE, as_derivs=TRUE)
+#' dz3<- Pfit(r, metrics = c("dz.dx", "dz.dy"))
 #' dirslp1<- DirSlp(alpha = 45, dz.dx= dz1$dz.dx, dz.dy= dz1$dz.dy)
 #' dirslp2<- DirSlp(alpha = 45, dz.dx= dz2$zx, dz.dy= dz2$zy)
-#' @details dz.dx can be calculated at a specified scale via `SlpAsp`, `Qfit` (zx and zy), or from an existing layer calculated by another program.
+#' dirslp3<- DirSlp(alpha = 45, dz.dx= dz3$dz.dx, dz.dy= dz3$dz.dy)
+#' @details dz.dx and dz.dy can be calculated at a specified scale via `SlpAsp`, `Pfit`, `Qfit` (zx and zy), or from an existing layer calculated by another program.
 #' @references
 #' Neteler, M., & Mitasova, H. (2008). Open source GIS: A GRASS GIS approach (3rd ed.). Springer.
 #' @import terra
