+++ R/RIE.R
@@ -10,9 +10,7 @@
 #' @param wopt list with named options for writing files as in writeRaster
 #' @return a SpatRaster or RasterLayer
 #' @examples 
-#' r<- rast(volcano, extent= ext(2667400, 2667400 + 
-#' ncol(volcano)*10, 6478700, 6478700 + nrow(volcano)*10), 
-#' crs = "EPSG:27200")
+#' r<- erupt()
 #' rie<- RIE(r, w=c(5,5), na.rm = TRUE)
 #' plot(rie)
 #' @import terra
