+++ R/RcppExports.R
@@ -1,58 +1,6 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-subset_mat_rows <- function(Input_Matrix, Input_Log_Vec) {
-    .Call(`_MultiscaleDTM_subset_mat_rows`, Input_Matrix, Input_Log_Vec)
-}
-
-C_OLS_params <- function(X, Y) {
-    .Call(`_MultiscaleDTM_C_OLS_params`, X, Y)
-}
-
-C_OLS_params2 <- function(Xt, XtX_inv, Y) {
-    .Call(`_MultiscaleDTM_C_OLS_params2`, Xt, XtX_inv, Y)
-}
-
-C_OLS_resid <- function(X, Y) {
-    .Call(`_MultiscaleDTM_C_OLS_resid`, X, Y)
-}
-
-C_OLS_resid2 <- function(X, Xt, XtX_inv, Y) {
-    .Call(`_MultiscaleDTM_C_OLS_resid2`, X, Xt, XtX_inv, Y)
-}
-
-C_Qfit1_narmT <- function(z, X_full, ni, nw) {
-    .Call(`_MultiscaleDTM_C_Qfit1_narmT`, z, X_full, ni, nw)
-}
-
-C_Qfit1_narmT_arma <- function(z, X_full, ni, nw) {
-    .Call(`_MultiscaleDTM_C_Qfit1_narmT_arma`, z, X_full, ni, nw)
-}
-
-C_Qfit1_narmF <- function(z, X, Xt, XtX_inv, ni, nw) {
-    .Call(`_MultiscaleDTM_C_Qfit1_narmF`, z, X, Xt, XtX_inv, ni, nw)
-}
-
-C_Qfit1_narmF_arma <- function(z, X, Xt, XtX_inv, ni, nw) {
-    .Call(`_MultiscaleDTM_C_Qfit1_narmF_arma`, z, X, Xt, XtX_inv, ni, nw)
-}
-
-C_Qfit2_narmT <- function(z, X_full, ni, nw) {
-    .Call(`_MultiscaleDTM_C_Qfit2_narmT`, z, X_full, ni, nw)
-}
-
-C_Qfit2_narmT_arma <- function(z, X_full, ni, nw) {
-    .Call(`_MultiscaleDTM_C_Qfit2_narmT_arma`, z, X_full, ni, nw)
-}
-
-C_Qfit2_narmF <- function(z, X, Xt, XtX_inv, ni, nw) {
-    .Call(`_MultiscaleDTM_C_Qfit2_narmF`, z, X, Xt, XtX_inv, ni, nw)
-}
-
-C_Qfit2_narmF_arma <- function(z, X, Xt, XtX_inv, ni, nw) {
-    .Call(`_MultiscaleDTM_C_Qfit2_narmF_arma`, z, X, Xt, XtX_inv, ni, nw)
-}
-
 C_AdjSD_narmT <- function(z, X_full, ni, nw) {
     .Call(`_MultiscaleDTM_C_AdjSD_narmT`, z, X_full, ni, nw)
 }
@@ -61,16 +9,24 @@ C_AdjSD_narmF <- function(z, X, Xt, XtX_inv, ni, nw) {
     .Call(`_MultiscaleDTM_C_AdjSD_narmF`, z, X, Xt, XtX_inv, ni, nw)
 }
 
-C_AdjSD_narmT_arma <- function(z, X_full, ni, nw) {
-    .Call(`_MultiscaleDTM_C_AdjSD_narmT_arma`, z, X_full, ni, nw)
+C_CountVals <- function(z, ni, nw) {
+    .Call(`_MultiscaleDTM_C_CountVals`, z, ni, nw)
 }
 
-C_AdjSD_narmF_arma <- function(z, X, Xt, XtX_inv, ni, nw) {
-    .Call(`_MultiscaleDTM_C_AdjSD_narmF_arma`, z, X, Xt, XtX_inv, ni, nw)
+C_Qfit1_narmT <- function(z, X_full, return_intercept, ni, nw) {
+    .Call(`_MultiscaleDTM_C_Qfit1_narmT`, z, X_full, return_intercept, ni, nw)
 }
 
-C_Pfit1_narmF <- function(z, X, Xt, XtX_inv, idx, ni, nw) {
-    .Call(`_MultiscaleDTM_C_Pfit1_narmF`, z, X, Xt, XtX_inv, idx, ni, nw)
+C_Qfit1_narmF <- function(z, X, Xt, XtX_inv, return_intercept, ni, nw) {
+    .Call(`_MultiscaleDTM_C_Qfit1_narmF`, z, X, Xt, XtX_inv, return_intercept, ni, nw)
+}
+
+C_Qfit2_narmT <- function(z, X_full, ni, nw) {
+    .Call(`_MultiscaleDTM_C_Qfit2_narmT`, z, X_full, ni, nw)
+}
+
+C_Qfit2_narmF <- function(z, X, Xt, XtX_inv, ni, nw) {
+    .Call(`_MultiscaleDTM_C_Qfit2_narmF`, z, X, Xt, XtX_inv, ni, nw)
 }
 
 C_TriArea <- function(a, b, c) {
@@ -79,17 +35,5 @@ C_TriArea <- function(a, b, c) {
 
 C_SurfaceArea <- function(z, x_res, y_res, na_rm, ni, nw) {
     .Call(`_MultiscaleDTM_C_SurfaceArea`, z, x_res, y_res, na_rm, ni, nw)
-}
-
-C_SurfaceArea_optimized <- function(z, x_res, y_res, na_rm, ni, nw) {
-    .Call(`_MultiscaleDTM_C_SurfaceArea_optimized`, z, x_res, y_res, na_rm, ni, nw)
-}
-
-C_CountVals <- function(z, ni, nw) {
-    .Call(`_MultiscaleDTM_C_CountVals`, z, ni, nw)
-}
-
-C_CountVals_optimized <- function(z, ni, nw) {
-    .Call(`_MultiscaleDTM_C_CountVals_optimized`, z, ni, nw)
 }
